requirements.txt
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
Deprecated==1.2.14
docker==7.1.0
exceptiongroup==1.2.2
google-api-core==2.20.0
google-api-python-client==2.148.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
httplib2==0.22.0
idna==3.10
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyGithub==2.4.0
PyJWT==2.9.0
PyNaCl==1.5.0
pyparsing==3.1.4
pytest==8.3.3
requests==2.32.3
rsa==4.9
tomli==2.0.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
wrapt==1.16.0
